+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 2.1.8
 
+- Fix for undefined currency exchange rate, add error message in the currency configuration page when an exchange rate could not be found (related #1751)
 - Fix the rounding of prices in the order product loop (related to #1753)
 - Add EQUAL to product loop filter by min or max (related to #1750)
 - Fix output value IS_DEFAULT in the product_sale_elements loop (related to #1745)